requirements.txt
#---------------------------------------------------------
# py -m pip install -r requirements.txt
#---------------------------------------------------------
requests
bs4
datetime
plexapi
cryptocode
ruamel.yaml
xmltodict
tqdm
scipy